Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 
 int main() {
-    float i;
-    scanf("%.4f",&i);
+    double i;
+    scanf("%f",&i);
 
-    printf("You entered: %.4f", i);
+    printf("You entered: %f", i);
     return 0;
 }
